refactor: update import paths for request and response models, and improve HTML formatting in login and register components
