+++ src/app/features/posts/pages/post-form/post-form.component.ts
@@ -5,8 +5,8 @@ import { CommonModule } from '@angular/common';
 import { PostService } from '../../../../core/services/post.service';
 import { ThemeService } from '../../../../core/services/theme.service';
 import { Theme } from '../../../../core/models/theme.model';
-import { CreatePostRequest } from '../../../../core/models/create-post-request.model';
-import { UpdatePostRequest } from '../../../../core/models/update-post-request.model';
+import { CreatePostRequest } from '../../../../core/models/request/create-post-request.model';
+import { UpdatePostRequest } from '../../../../core/models/request/update-post-request.model';
 import { HttpErrorResponse } from '@angular/common/http';
 import { AuthService } from '../../../../core/services/auth.service';
 import { ToastService } from '../../../../core/services/toast.service';
