+++ src/app/features/posts/pages/post-list/post-list.component.ts
@@ -2,7 +2,7 @@ import { Component, inject } from '@angular/core';
 import { CommonModule }      from '@angular/common';
 import { Observable }        from 'rxjs';
 import { PostService }       from '../../../../core/services/post.service';
-import { PostResponse }      from '../../../../core/models/post-reponse.model';
+import { PostResponse }      from '../../../../core/models/response/post-reponse.model';
 import { Router } from 'express';
 import { RouterModule } from '@angular/router';
 import { PostCardComponent } from '../../components/post-card/post-card.component';
